interactive_scene/scene_builder/requirements.txt
certifi==2020.11.8
cycler==0.10.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.3.2
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pycairo==1.20.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.3
six==1.15.0
texttable==1.6.3
pymeshlab>=2021.07